pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "ForensIQ Enterprise Digital Forensics Platform"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "werkzeug>=3.0.0",
    "reportlab>=4.0.0",
    "matplotlib>=3.7.0",
    "psutil>=5.9.0",
    "pyotp>=2.8.0",
    "qrcode>=7.4.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "python-magic>=0.4.27",
    "email-validator>=2.2.0",
]
